requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pyperclip>=1.8.2
tkinterdnd2>=0.3.0
pystray>=0.19.0
